Fix: Fix independent clipper model
