+++ src/app/models/ThreeBaseModel.ts
@@ -1,5 +1,6 @@
 import * as THREE from 'three';
 import { v4 as uuid } from 'uuid';
+import EventEmitter from 'events';
 import { HEAD_PRINTING } from '../constants';
 import { SvgModelElement } from './BaseModel';
 import type ModelGroup from './ModelGroup';
@@ -83,13 +84,14 @@ const DEFAULT_TRANSFORMATION: ModelTransformation = {
 
 // BaseModel only do data process
 // isolated from Model.js which renamed to ThreeModel.js
-export default class BaseModel {
+export default class BaseModel extends EventEmitter {
     public headType: typeof HEAD_PRINTING = HEAD_PRINTING;
     public sourceType: '3d' = '3d';
 
     public modelID: string;
     public originModelID: string;
     public modelName: string;
+    // public visible: boolean;
     public sourceHeight: number;
     public sourceWidth: number;
     public originalName: string;
@@ -122,6 +124,7 @@ export default class BaseModel {
     protected modelModeMaterial: THREE.MeshStandardMaterial = new THREE.MeshStandardMaterial({ color: 0xe0e0e0, visible: false });
 
     public constructor(modelInfo: ModelInfo, modelGroup: ModelGroup) {
+        super();
         this.modelGroup = modelGroup;
 
         Object.keys(modelInfo).forEach((key) => {
