+++ src/app/workers/calaClippingSkin.ts
@@ -3,6 +3,8 @@ import { polyDiff, polyIntersection, polyOffset } from '../../shared/lib/clipper
 
 type TPoint = { x: number, y: number, z?: number }
 
+type TPolygon = TPoint[][]
+
 const intersectionSkin = (subPaths, vectorsArray) => {
     if (!subPaths || subPaths.length === 0 || !vectorsArray || vectorsArray.length === 0) {
         return [];
@@ -22,14 +24,39 @@ const intersectionSkin = (subPaths, vectorsArray) => {
 };
 
 type TMessage = {
-    index: string, currentInnerWall: TPoint[], otherLayers: TPoint[][], lineWidth: number
+    index: string, currentInnerWall: TPolygon[], otherLayers: TPolygon[][], lineWidth: number
 }
+
+// const unionPolygons = (polygons: TPolygon[], area?: TPolygon) => {
+//     if (polygons.length === 0) {
+//         return area;
+//     }
+//     let polygon = polygons.shift();
+//     if (!area) {
+//         area = polygon;
+//         polygon = polygons.shift();
+
+//         if (polygons.length === 0) {
+//             return area;
+//         }
+//     }
+//     const _area = polyUnion(area, [polygon]);
+//     return unionPolygons(polygons, _area);
+// };
+
+const unionPolygons = (polygons: TPolygon[]): TPolygon => {
+    return polygons.reduce((p, c) => {
+        p.push(...c);
+        return p;
+    }, []);
+};
 
 const calaClippingSkin = ({ currentInnerWall, otherLayers, lineWidth }: TMessage) => {
     return new Observable((observer) => {
         const commonArea = otherLayers.reduce((p, c) => {
-            return intersectionSkin(c, p);
-        }, otherLayers[0] || []);
+            return intersectionSkin(unionPolygons(c), p);
+        }, otherLayers[0] ? unionPolygons(otherLayers[0]) : []);
+
 
         let skin;
         let infill;
@@ -38,11 +65,14 @@ const calaClippingSkin = ({ currentInnerWall, otherLayers, lineWidth }: TMessage
             skin = currentInnerWall;
             infill = [];
         } else {
-            skin = polyDiff(currentInnerWall, commonArea);
-            infill = commonArea;
+            skin = polyDiff(unionPolygons(currentInnerWall), commonArea);
+            infill = [commonArea];
+
+            skin = [skin];
         }
         if (skin && skin.length > 0) {
-            skin = polyOffset(skin, -lineWidth);
+            skin = polyOffset(unionPolygons(skin), -lineWidth);
+            skin = [skin];
         }
         observer.next({
             infill,
